# storage/api-docs/art-info-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Art-UA Info API",
        "description": "API для інформаційного порталу art-ua.info — новини, події, статті про мистецтво",
        "contact": {
            "name": "Art-UA Info Support",
            "email": "info@art-ua.info"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://art-ua.info/api/v1",
            "description": "Production Server"
        },
        {
            "url": "https://save-art.ddev.site/api/v1",
            "description": "Local Development"
        }
    ],
    "paths": {
...
        "/v1/artists": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Список митців",
                "description": "Повертає список митців (користувачів, які мають опубліковані проєкти)",
                "operationId": "getArtists",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Пошук по імені митця",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Іван"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Кількість на сторінку (макс. 50)",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 50
                        },
                        "example": 20
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер сторінки",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "example": 1
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Мова відповіді (uk, en). Якщо не вказано — повертає об'єкт з усіма мовами",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "uk",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список митців",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Artist"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/artists/{slug}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Профіль митця",
                "description": "Повертає публічний профіль митця з його даними та соціальними мережами",
                "operationId": "getArtist",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Унікальний slug митця",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ivan-franko-abc123"
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Мова відповіді (uk, en). Якщо не вказано — повертає об'єкт з усіма мовами",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "uk",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Профіль митця",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Artist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Митця не знайдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/artists/{slug}/projects": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Проєкти митця",
                "description": "Повертає список публічних проєктів конкретного митця",
                "operationId": "getArtistProjects",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug митця",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ivan-franko-abc123"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фільтр по статусу проєкту",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "announced",
                                "in_progress",
                                "completed",
                                "sold"
                            ]
                        },
                        "example": "announced"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Кількість на сторінку (макс. 50)",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 50
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Мова відповіді (uk, en). Якщо не вказано — повертає об'єкт з усіма мовами",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "uk",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список проєктів митця",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProjectList"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Митця не знайдено"
                    }
                }
            }
        },
        "/v1/artists/{slug}/photos": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Портфоліо митця",
                "description": "Повертає список фото з портфоліо митця",
                "operationId": "getArtistPhotos",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug митця",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список фото",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserPhoto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Митця не знайдено"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "LocalizedString": {
                "title": "LocalizedString",
                "description": "Мультимовний текст (uk/en)",
                "properties": {
                    "uk": {
                        "type": "string"
                    },
                    "en": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "PaginationMeta",
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "title": "PaginationLinks",
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "ErrorResponse",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Author": {
                "title": "Author",
                "description": "Автор проєкту",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Іван Франко"
                    },
                    "slug": {
                        "type": "string",
                        "example": "ivan-franko",
                        "nullable": true
                    },
                    "avatar_url": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Artist": {
                "title": "Artist",
                "description": "Публічний профіль митця",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Іван Франко"
                    },
                    "slug": {
                        "type": "string",
                        "example": "ivan-franko"
                    },
                    "avatar_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "profession": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/LocalizedString"
                            }
                        ],
                        "nullable": true
                    },
                    "bio": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/LocalizedString"
                            }
                        ],
                        "nullable": true
                    },
                    "city": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/LocalizedString"
                            }
                        ],
                        "nullable": true
                    },
                    "country": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/LocalizedString"
                            }
                        ],
                        "nullable": true
                    },
                    "social": {
                        "properties": {
                            "website": {
                                "type": "string",
                                "nullable": true
                            },
                            "facebook": {
                                "type": "string",
                                "nullable": true
                            },
                            "instagram": {
                                "type": "string",
                                "nullable": true
                            },
                            "youtube": {
                                "type": "string",
                                "nullable": true
                            },
                            "linkedin": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "projects_count": {
                        "type": "integer",
                        "example": 5
                    },
                    "completed_projects_count": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ProjectList": {
                "title": "ProjectList",
                "description": "Проєкт у списку (скорочена інформація)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "slug": {
                        "type": "string",
                        "example": "miy-noviy-proekt-abc123"
                    },
                    "status": {
                        "type": "string",
                        "example": "announced"
                    },
                    "status_label": {
                        "type": "string",
                        "example": "Оголошений"
                    },
                    "status_moderation": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "approved",
                            "rejected"
                        ],
                        "example": "pending",
                        "nullable": true
                    },
                    "title": {
                        "$ref": "#/components/schemas/LocalizedString"
                    },
                    "short_description": {
                        "$ref": "#/components/schemas/LocalizedString"
                    },
                    "cover_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "art_category": {
                        "type": "string",
                        "example": "visual"
                    },
                    "art_category_label": {
                        "type": "string",
                        "example": "Візуальне мистецтво"
                    },
                    "currency": {
                        "type": "string",
                        "example": "UAH"
                    },
                    "budget_goal": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "budget_collected": {
                        "type": "number",
                        "format": "float",
                        "example": 12500
                    },
                    "progress_percentage": {
                        "type": "number",
                        "format": "float",
                        "example": 25
                    },
                    "estimated_days": {
                        "description": "Орієнтовна кількість днів на реалізацію",
                        "type": "integer",
                        "example": 90,
                        "nullable": true
                    },
                    "days_left": {
                        "description": "Залишилось днів до планованого завершення",
                        "type": "integer",
                        "example": 30,
                        "nullable": true
                    },
                    "likes_count": {
                        "type": "integer",
                        "example": 42
                    },
                    "donors_count": {
                        "type": "integer",
                        "example": 15
                    },
                    "tags": {
                        "description": "Теги проєкту",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "живопис",
                            "сучасне мистецтво"
                        ],
                        "nullable": true
                    },
                    "parameters": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProjectParameter"
                        }
                    },
                    "author": {
                        "$ref": "#/components/schemas/Author"
                    },
                    "announced_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "planned_completion_at": {
                        "description": "Планова дата завершення",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ProjectParameter": {
                "title": "ProjectParameter",
                "description": "Значення характеристики проєкту (включно з незаповненими — value: null)",
                "properties": {
                    "parameter_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "parameter": {
                        "description": "Назва характеристики. Якщо передано language — строка, інакше об'єкт {uk, en}",
                        "example": "Жанр"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "list",
                            "custom"
                        ],
                        "example": "list"
                    },
                    "value_id": {
                        "type": "integer",
                        "example": 12,
                        "nullable": true
                    },
                    "value": {
                        "description": "Значення. Якщо передано language — строка, інакше об'єкт {uk, en} (для custom) або строка (для list)",
                        "example": "Роман",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UserPhoto": {
                "title": "UserPhoto",
                "description": "Фото з портфоліо автора (митця чи організації)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "image_url": {
                        "type": "string"
                    },
                    "likes_count": {
                        "type": "integer",
                        "example": 12
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token. Формат: Bearer <token>",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "apiKey": {
                "type": "apiKey",
                "description": "API ключ для доступу до API. Передається у заголовку X-Api-Key.",
                "name": "X-Api-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKey": []
        }
    ]
}